chore: Reorganize some modules
